requirement.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-extensions==3.2.3
djangorestframework==3.15.2
graphviz==0.20.3
idna==3.10
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
urllib3==2.2.3
